app/core/auth.py
# auth.py
from datetime import datetime, timedelta
from jose import JWTError, jwt
from passlib.context import CryptContext
from database.user import *

from os import getenv

SECRET_KEY = getenv("API_SECRET_KEY")
ALGORITHM = getenv("API_ALGORITHM")
ACCESS_TOKEN_EXPIRE_MINUTES = int(getenv("API_ACCESS_TOKEN_EXPIRE_MINUTES"))

# Contexto para hash de senha
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def authenticate_user(username: str, password: str):
    user = get_user_db(username)
    if not user or not verify_password(password, user["password"]):
        return None  # Usuário ou senha inválidos
    return user


def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def decode_access_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
