requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
boto3==1.35.49
botocore==1.35.49
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
DateTime==5.5
ecdsa==0.19.0
fastapi==0.115.3
h11==0.14.0
idna==3.10
jmespath==1.0.1
paho-mqtt==2.1.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
rsa==4.9
s3transfer==0.10.3
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
uvicorn==0.32.0
zope.interface==7.1.1
